refactor: move topic related code to topics.go
