+++ pkg/messaging/topics.go
@@ -4,11 +4,17 @@
 package messaging
 
 import (
+	"context"
 	"fmt"
 	"net/url"
 	"strings"
 
+	grpcChannelsV1 "github.com/absmach/supermq/api/grpc/channels/v1"
+	grpcCommonV1 "github.com/absmach/supermq/api/grpc/common/v1"
+	grpcDomainsV1 "github.com/absmach/supermq/api/grpc/domains/v1"
 	"github.com/absmach/supermq/pkg/errors"
+	"github.com/dgraph-io/ristretto/v2"
+	"github.com/gofrs/uuid/v5"
 )
 
 const (
@@ -17,14 +23,217 @@ const (
 )
 
 var (
-	ErrMalformedTopic    = errors.New("malformed topic")
-	ErrMalformedSubtopic = errors.New("malformed subtopic")
 	mqWildcards          = "+#"
 	wildcards            = "*>"
 	subtopicInvalidChars = " #+"
 	wildcardsReplacer    = strings.NewReplacer("+", "*", "#", ">")
 	pathReplacer         = strings.NewReplacer("/", ".")
+
+	DefaultCacheConfig = CacheConfig{
+		NumCounters: 2e5,     // 200k
+		MaxCost:     1 << 20, // 1MB
+		BufferItems: 64,
+	}
+
+	ErrMalformedTopic       = errors.New("malformed topic")
+	ErrMalformedSubtopic    = errors.New("malformed subtopic")
+	ErrEmptyRouteID         = errors.New("empty route or id")
+	ErrFailedResolveDomain  = errors.New("failed to resolve domain route")
+	ErrFailedResolveChannel = errors.New("failed to resolve channel route")
+	ErrCreateCache          = errors.New("failed to create cache")
 )
+
+type CacheConfig struct {
+	NumCounters int64 `env:"NUM_COUNTERS" envDefault:"200000"`  // number of keys to track frequency of.
+	MaxCost     int64 `env:"MAX_COST"     envDefault:"1048576"` // maximum cost of cache.
+	BufferItems int64 `env:"BUFFER_ITEMS" envDefault:"64"`      // number of keys per Get buffer.
+}
+
+type parsedTopic struct {
+	domainID  string
+	channelID string
+	subtopic  string
+	err       error
+}
+
+// TopicParser defines methods for parsing publish and subscribe topics.
+// It uses a cache to store parsed topics for quick retrieval.
+// It also resolves domain and channel IDs if requested.
+type TopicParser interface {
+	ParsePublishTopic(ctx context.Context, topic string, resolve bool) (domainID, channelID, subtopic string, err error)
+	ParseSubscribeTopic(ctx context.Context, topic string, resolve bool) (domainID, channelID, subtopic string, err error)
+}
+
+type parser struct {
+	resolver TopicResolver
+	cache    *ristretto.Cache[string, *parsedTopic]
+}
+
+// NewTopicParser creates a new instance of TopicParser.
+func NewTopicParser(cfg CacheConfig, channels grpcChannelsV1.ChannelsServiceClient, domains grpcDomainsV1.DomainsServiceClient) (TopicParser, error) {
+	cache, err := ristretto.NewCache(&ristretto.Config[string, *parsedTopic]{
+		NumCounters: cfg.NumCounters,
+		MaxCost:     cfg.MaxCost,
+		BufferItems: cfg.BufferItems,
+		Cost:        costFunc,
+	})
+	if err != nil {
+		return nil, errors.Wrap(ErrCreateCache, err)
+	}
+	return &parser{
+		cache:    cache,
+		resolver: NewTopicResolver(channels, domains),
+	}, nil
+}
+
+func (p *parser) ParsePublishTopic(ctx context.Context, topic string, resolve bool) (string, string, string, error) {
+	val, ok := p.cache.Get(topic)
+	if ok {
+		return val.domainID, val.channelID, val.subtopic, val.err
+	}
+	domainID, channelID, subtopic, err := ParsePublishTopic(topic)
+	if err != nil {
+		p.saveToCache(topic, "", "", "", err)
+		return "", "", "", err
+	}
+	var isRoute bool
+	if resolve {
+		domainID, channelID, isRoute, err = p.resolver.Resolve(ctx, domainID, channelID)
+		if err != nil {
+			return "", "", "", err
+		}
+	}
+	if !isRoute {
+		p.saveToCache(topic, domainID, channelID, subtopic, nil)
+	}
+
+	return domainID, channelID, subtopic, nil
+}
+
+func (p *parser) ParseSubscribeTopic(ctx context.Context, topic string, resolve bool) (string, string, string, error) {
+	domainID, channelID, subtopic, err := ParseSubscribeTopic(topic)
+	if err != nil {
+		return "", "", "", err
+	}
+	if resolve {
+		domainID, channelID, _, err = p.resolver.Resolve(ctx, domainID, channelID)
+		if err != nil {
+			return "", "", "", err
+		}
+	}
+
+	return domainID, channelID, subtopic, nil
+}
+
+func (p *parser) saveToCache(topic string, domainID, channelID, subtopic string, err error) {
+	p.cache.Set(topic, &parsedTopic{
+		domainID:  domainID,
+		channelID: channelID,
+		subtopic:  subtopic,
+		err:       err,
+	}, 0)
+}
+
+func costFunc(val *parsedTopic) int64 {
+	errLen := 0
+	if val.err != nil {
+		errLen = len(val.err.Error())
+	}
+	cost := int64(len(val.domainID) + len(val.channelID) + len(val.subtopic) + errLen)
+
+	return cost
+}
+
+// TopicResolver contains definitions for resolving domain and channel IDs
+// from their respective routes from the message topic.
+type TopicResolver interface {
+	Resolve(ctx context.Context, domain, channel string) (domainID string, channelID string, isRoute bool, err error)
+	ResolveTopic(ctx context.Context, topic string) (rtopic string, err error)
+}
+
+type resolver struct {
+	channels grpcChannelsV1.ChannelsServiceClient
+	domains  grpcDomainsV1.DomainsServiceClient
+}
+
+// NewTopicResolver creates a new instance of TopicResolver.
+func NewTopicResolver(channelsClient grpcChannelsV1.ChannelsServiceClient, domainsClient grpcDomainsV1.DomainsServiceClient) TopicResolver {
+	return &resolver{
+		channels: channelsClient,
+		domains:  domainsClient,
+	}
+}
+
+func (r *resolver) Resolve(ctx context.Context, domain, channel string) (string, string, bool, error) {
+	if domain == "" || channel == "" {
+		return "", "", false, ErrEmptyRouteID
+	}
+
+	domainID, isdomainRoute, err := r.resolveDomain(ctx, domain)
+	if err != nil {
+		return "", "", false, errors.Wrap(ErrFailedResolveDomain, err)
+	}
+	channelID, isChannelRoute, err := r.resolveChannel(ctx, channel, domainID)
+	if err != nil {
+		return "", "", false, errors.Wrap(ErrFailedResolveChannel, err)
+	}
+	isRoute := isdomainRoute || isChannelRoute
+
+	return domainID, channelID, isRoute, nil
+}
+
+func (r *resolver) ResolveTopic(ctx context.Context, topic string) (string, error) {
+	domain, channel, subtopic, err := ParseTopic(topic)
+	if err != nil {
+		return "", errors.Wrap(ErrMalformedTopic, err)
+	}
+
+	domainID, channelID, _, err := r.Resolve(ctx, domain, channel)
+	if err != nil {
+		return "", err
+	}
+	rtopic := EncodeAdapterTopic(domainID, channelID, subtopic)
+
+	return rtopic, nil
+}
+
+func (r *resolver) resolveDomain(ctx context.Context, domain string) (string, bool, error) {
+	if validateUUID(domain) == nil {
+		return domain, false, nil
+	}
+	d, err := r.domains.RetrieveByRoute(ctx, &grpcCommonV1.RetrieveByRouteReq{
+		Route: domain,
+	})
+	if err != nil {
+		return "", false, err
+	}
+
+	return d.Entity.Id, true, nil
+}
+
+func (r *resolver) resolveChannel(ctx context.Context, channel, domainID string) (string, bool, error) {
+	if validateUUID(channel) == nil {
+		return channel, false, nil
+	}
+	c, err := r.channels.RetrieveByRoute(ctx, &grpcCommonV1.RetrieveByRouteReq{
+		Route:    channel,
+		DomainId: domainID,
+	})
+	if err != nil {
+		return "", false, err
+	}
+
+	return c.Entity.Id, true, nil
+}
+
+func validateUUID(extID string) (err error) {
+	id, err := uuid.FromString(extID)
+	if id.String() != extID || err != nil {
+		return err
+	}
+
+	return nil
+}
 
 func ParsePublishTopic(topic string) (domainID, chanID, subtopic string, err error) {
 	domainID, chanID, subtopic, err = ParseTopic(topic)
